+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneDeviceRemediation/MSFT_IntuneDeviceRemediation.psm1
@@ -228,18 +228,22 @@ function Get-TargetResource
             AccessTokens                = $AccessTokens
             #endregion
         }
+
         $assignmentsValues = Get-MgBetaDeviceManagementDeviceHealthScriptAssignment -DeviceHealthScriptId $Id
         $assignmentResult = @()
-        foreach ($assignmentEntry in $AssignmentsValues)
+        foreach ($assignment in $assignmentsValues)
         {
-            $assignmentValue = @{
-                dataType = $assignmentEntry.Target.AdditionalProperties.'@odata.type'
-                deviceAndAppManagementAssignmentFilterType = $(if ($null -ne $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterType)
-                    {$assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterType.ToString()})
-                deviceAndAppManagementAssignmentFilterId = $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterId
-                groupId = $assignmentEntry.Target.AdditionalProperties.groupId
+            $assignmentResult += @{
+                RunRemediationScript = $assignment.RunRemediationScript
+                RunSchedule = @{
+                    DataType = $assignment.RunSchedule.AdditionalProperties.'@odata.type'
+                    Date = $assignment.RunSchedule.AdditionalProperties.date
+                    Interval = $assignment.RunSchedule.Interval
+                    Time = Get-Date -Format 'HH:mm:ss' -Date $assignment.RunSchedule.AdditionalProperties.time
+                    UseUtc = $assignment.RunSchedule.AdditionalProperties.useUtc
+                }
+                Assignment = ConvertFrom-IntunePolicyAssignment -IncludeDeviceFilter:$true -Assignment $assignment
             }
-            $assignmentResult += $assignmentValue
         }
         $results.Add('Assignments', $assignmentResult)
 
@@ -397,14 +401,43 @@ function Set-TargetResource
         $assignmentsHash = @()
         foreach ($assignment in $Assignments)
         {
-            $assignmentsHash += Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignment
+            $assignmentTarget = ConvertTo-IntunePolicyAssignment -Assignments $assignment.Assignment
+            $runSchedule = $null
+            if ($null -ne $assignment.RunSchedule.DataType) {
+                $runSchedule = @{
+                    '@odata.type' = $assignment.RunSchedule.DataType
+                }
+                if (-not [string]::IsNullOrEmpty($assignment.RunSchedule.Date))
+                {
+                    $runSchedule.Add('date', $assignment.RunSchedule.Date)
+                }
+                if (-not [string]::IsNullOrEmpty($assignment.RunSchedule.Interval))
+                {
+                    $runSchedule.Add('interval', $assignment.RunSchedule.Interval)
+                }
+                if (-not [string]::IsNullOrEmpty($assignment.RunSchedule.Time))
+                {
+                    $runSchedule.Add('time', $assignment.RunSchedule.Time)
+                }
+                if (-not [string]::IsNullOrEmpty($assignment.RunSchedule.UseUtc))
+                {
+                    $runSchedule.Add('useUtc', $assignment.RunSchedule.UseUtc)
+                }
+            }
+            $assignmentsHash += @{
+                runRemediationScript = $assignment.RunRemediationScript
+                runSchedule = $runSchedule
+                target = $assignmentTarget.target
+            }
         }
 
-        if ($policy.id)
+        if ($policy.Id)
         {
-            Update-DeviceConfigurationPolicyAssignment -DeviceConfigurationPolicyId $policy.id `
-                -Targets $assignmentsHash `
-                -Repository 'deviceManagement/deviceHealthScripts'
+            $uri = "/beta/deviceManagement/deviceHealthScripts/$($policy.Id)/assign"
+            $body = @{
+                deviceHealthScriptAssignments = $assignmentsHash
+            } | ConvertTo-Json -Depth 20
+            Invoke-MgGraphRequest -Method POST -Uri $uri -Body $body -ErrorAction Stop 4> $null
         }
         #endregion
     }
@@ -433,20 +466,49 @@ function Set-TargetResource
         Update-MgBetaDeviceManagementDeviceHealthScript  `
             -DeviceHealthScriptId $currentInstance.Id `
             -BodyParameter $UpdateParameters
+
         $assignmentsHash = @()
         foreach ($assignment in $Assignments)
         {
-            $assignmentsHash += Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignment
+            $assignmentTarget = ConvertTo-IntunePolicyAssignment -Assignments $assignment.Assignment
+            $runSchedule = $null
+            if ($null -ne $assignment.RunSchedule.DataType) {
+                $runSchedule = @{
+                    '@odata.type' = $assignment.RunSchedule.DataType
+                }
+                if (-not [string]::IsNullOrEmpty($assignment.RunSchedule.Date))
+                {
+                    $runSchedule.Add('date', $assignment.RunSchedule.Date)
+                }
+                if (-not [string]::IsNullOrEmpty($assignment.RunSchedule.Interval))
+                {
+                    $runSchedule.Add('interval', $assignment.RunSchedule.Interval)
+                }
+                if (-not [string]::IsNullOrEmpty($assignment.RunSchedule.Time))
+                {
+                    $runSchedule.Add('time', $assignment.RunSchedule.Time)
+                }
+                if (-not [string]::IsNullOrEmpty($assignment.RunSchedule.UseUtc))
+                {
+                    $runSchedule.Add('useUtc', $assignment.RunSchedule.UseUtc)
+                }
+            }
+            $assignmentsHash += @{
+                runRemediationScript = $assignment.RunRemediationScript
+                runSchedule = $runSchedule
+                target = $assignmentTarget.target
+            }
         }
-        Update-DeviceConfigurationPolicyAssignment `
-            -DeviceConfigurationPolicyId $currentInstance.id `
-            -Targets $assignmentsHash `
-            -Repository 'deviceManagement/deviceHealthScripts'
+        $uri = "/beta/deviceManagement/deviceHealthScripts/$($currentInstance.Id)/assign"
+        $body = @{
+            deviceHealthScriptAssignments = $assignmentsHash
+        } | ConvertTo-Json -Depth 20
+        Invoke-MgGraphRequest -Method POST -Uri $uri -Body $body -ErrorAction Stop 4> $null
         #endregion
     }
     elseif ($Ensure -eq 'Absent' -and $currentInstance.Ensure -eq 'Present')
     {
-        Write-Verbose -Message "Removing the Intune Device Remediation with Id {$($currentInstance.Id)}" 
+        Write-Verbose -Message "Removing the Intune Device Remediation with Id {$($currentInstance.Id)}"
         #region resource generator code
         Remove-MgBetaDeviceManagementDeviceHealthScript -DeviceHealthScriptId $currentInstance.Id
         #endregion
@@ -582,17 +644,39 @@ function Test-TargetResource
     {
         $source = $PSBoundParameters.$key
         $target = $CurrentValues.$key
-        if ($source.getType().Name -like '*CimInstance*')
+        if ($null -ne $source -and $source.GetType().Name -like '*CimInstance*')
         {
             $source = Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $source
 
-            $testResult = Compare-M365DSCComplexObject `
-                -Source ($source) `
-                -Target ($target)
-
-            if (-Not $testResult)
+            if ($key -eq "Assignments")
             {
-                $testResult = $false
+                [hashtable[]]$sourcesWithoutAssignment = @()
+                foreach ($sourceObject in $source)
+                {
+                    $sourceWithoutAssignment = $sourceObject.Clone()
+                    $sourcesWithoutAssignment += $sourceWithoutAssignment
+                }
+
+                [hashtable[]]$targetsWithoutAssignment = @()
+                foreach ($targetObject in $target)
+                {
+                    $targetWithoutAssignment = $targetObject.Clone()
+                    $targetsWithoutAssignment += $targetWithoutAssignment
+                }
+
+                $testResult = Compare-M365DSCComplexObject `
+                    -Source ($sourcesWithoutAssignment) `
+                    -Target ($targetsWithoutAssignment)
+            }
+            else
+            {
+                $testResult = Compare-M365DSCComplexObject `
+                    -Source ($source) `
+                    -Target ($target)
+            }
+
+            if (-not $testResult)
+            {
                 break
             }
 
@@ -600,11 +684,8 @@ function Test-TargetResource
         }
     }
 
-    $ValuesToCheck.remove('Id') | Out-Null
-    $ValuesToCheck.Remove('Credential') | Out-Null
-    $ValuesToCheck.Remove('ApplicationId') | Out-Null
-    $ValuesToCheck.Remove('TenantId') | Out-Null
-    $ValuesToCheck.Remove('ApplicationSecret') | Out-Null
+    $ValuesToCheck.Remove('Id') | Out-Null
+    $ValuesToCheck = Remove-M365DSCAuthenticationParameter -BoundParameters $ValuesToCheck
 
     Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
     Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $ValuesToCheck)"
@@ -752,7 +833,15 @@ function Export-TargetResource
             }
             if ($Results.Assignments)
             {
-                $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString -ComplexObject $Results.Assignments -CIMInstanceName DeviceManagementConfigurationPolicyAssignments
+                foreach ($assignment in $Results.Assignments)
+                {
+                    $runSchedule = Get-M365DSCDRGComplexTypeToString -ComplexObject $assignment.RunSchedule -CIMInstanceName MSFT_IntuneDeviceRemediationRunSchedule
+                    $assignment.RunSchedule = $runSchedule
+                    $target = Get-M365DSCDRGComplexTypeToString -ComplexObject $assignment.Assignment -CIMInstanceName MSFT_DeviceManagementConfigurationPolicyAssignments
+                    $assignment.Assignment = $target | Select-Object -First 1
+                    $assignment = Get-M365DSCDRGComplexTypeToString -ComplexObject $assignment -CIMInstanceName MSFT_IntuneDeviceRemediationPolicyAssignments
+                }
+                $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString -ComplexObject $Results.Assignments -CIMInstanceName MSFT_IntuneDeviceRemediationPolicyAssignments
                 if ($complexTypeStringResult)
                 {
                     $Results.Assignments = $complexTypeStringResult
@@ -777,7 +866,10 @@ function Export-TargetResource
             }
             if ($Results.Assignments)
             {
-                $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "Assignments" -isCIMArray:$true
+                $currentDSCBlock = (Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "Assignments" -isCIMArray:$true).Replace("''", "'")
+                $currentDSCBlock = [Regex]::Replace($currentDSCBlock, "Assignment = '\r\n                ", "Assignment = ")
+                $currentDSCBlock = $currentDSCBlock.Replace("RunSchedule = '", "RunSchedule = ").Replace("}'", "}")
+                $currentDSCBlock = [Regex]::Replace($currentDSCBlock, "\r\n            '", "")
             }
 
             $dscContent += $currentDSCBlock